sim_efc_pwp_demo not working correctly
