+++ lina/__init__.py
@@ -1,5 +1,5 @@
-from . import math_module
-from . import utils, shmim_utils, coro_utils, efc, iefc, aefc, llowfsc, dm 
+from . import math_module, rt_utils
+from . import utils, coro_utils, efc, iefc, aefc, llowfsc, dm 
 
 __version__ = '0.1.0'
